Move `FarmerPieceGetter` into its own module
